Isolated plist-specific functionality into package org.catacombae.plist.
The plist parsing functionality was previously split between XMLNode,
which is supposed to belong to the XML parser and not deal with
specific applications of XML, and Plist.java in
org.catacombae.dmg.udif, which is specific to UDIF parsing.

To get a more general plist parsing facility, the org.catacombae.plist
package was created which isolates the plist parsing specifics (and
also introduces a possibility of adding support for non-XML plist
formats in the future (binary plists for instance).