Fix XmlPlistNode.getChildren() on CFArray (fix unsound#3)
Currently, XmlPlistNode.getChildren() only works for dicts, not for
arrays. However, in #3, this method gets called by Plist.getPartition()
on an CFArray of CFDicts of partitions. When that happens, the current
implementation of getChildren() returns an empty array, making the copy
operation do nothing.

This change makes getChildren on a CFArray return the lists of children
nodes. This seems to have no adverse effects elsewhere in my very
limited testing, but of course I have no guarantees.