Create a File/RandomAccessFile abstraction for sparsebundle interfaces.
This decouples the sparsebundle interfaces from the java.io.File and
java.io.RandomAccessFile constraints and allows us to use sparsebundle
interfaces with sparsebundle data that is not stored in the system's
Java-accessible file system.
It could be supplied over the network or loaded from a location embedded
in a non-mounted file system or image. It's up to the caller to
implement the necessary interfaces.

A default File/RandomAccessFile-backed implementation, JavaFileAccessor,
exists in SparseBundle.java.