DMGExtractor.java: Restructure extractProcedure to clean up on exit.
When we wanted out of extractProcedure we simply called
System.exit(int), which probably cleans up in a primitive way but it
doesn't call the proper 'close()' methods for all affected streams.

So we eliminate the System.exit(int) calls and use regular program flow
constructs, breaking out the main body of 'extractProcedure' to a new
function 'extractUDIF' (which makes sense anyway because
'extractProcedure' should be general, not UDIF-specific). When
'extractUDIF' (or 'copyData') returns, we clean up after it.