V2Header.java: Emit warnings for unexpected values of fields.
Some field members in the V2 header are not yet fully understood. Since
we don't know the exact meaning of these fields we now emit warning
messages on System.err when they deviate from what we have seen
previously.
This helps the developer to get valuable feedback for non-standard
images, and hopefully to understand the meaning of the fields.